+++ tests/api-resources/passages.test.ts
@@ -9,8 +9,8 @@ const client = new Letta({
 
 describe('resource passages', () => {
   // Prism tests are disabled
-  test.skip('search: only required params', async () => {
-    const responsePromise = client.passages.search({ query: 'query' });
+  test.skip('search', async () => {
+    const responsePromise = client.passages.search({});
     const rawResponse = await responsePromise.asResponse();
     expect(rawResponse).toBeInstanceOf(Response);
     const response = await responsePromise;
@@ -18,19 +18,5 @@ describe('resource passages', () => {
     const dataAndResponse = await responsePromise.withResponse();
     expect(dataAndResponse.data).toBe(response);
     expect(dataAndResponse.response).toBe(rawResponse);
-  });
-
-  // Prism tests are disabled
-  test.skip('search: required and optional params', async () => {
-    const response = await client.passages.search({
-      query: 'query',
-      agent_id: 'agent_id',
-      archive_id: 'archive_id',
-      end_date: '2019-12-27T18:11:19.117Z',
-      limit: 1,
-      start_date: '2019-12-27T18:11:19.117Z',
-      tag_match_mode: 'any',
-      tags: ['string'],
-    });
   });
 });